pyproject.toml
[project]
name = "stockmate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "black>=25.9.0",
    "fastapi[standard]>=0.117.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.14.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=app --cov-report=term-missing"

[tool.coverage.run]
branch = true
source = ["app"]
omit = [
    "tests/*",
    "app/db/models/*",
    ".venv/*",
    "*/__init__.py",
    "app/schemas/*",
    "app/clients/fmp/models/*",
]

[tool.ruff]
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    ".git",
    "migrations",
    "app/db/models/__init__.py",
    "alembic",
]
